Refactor Settings reader/writer into their own classes
