+++ InfiniteGoodsMod/InfiniteGoods.csproj
@@ -52,12 +52,14 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Goods.cs" />
-    <Compile Include="Settings.cs" />
     <Compile Include="GoodsMonitor.cs" />
     <Compile Include="ModIdentity.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="LoadingExtension.cs" />
-    <Compile Include="SettingsPanel.cs" />
+    <Compile Include="Settings\LegacySettingsFileReader.cs" />
+    <Compile Include="Settings\Settings.cs" />
+    <Compile Include="Settings\SettingsFileParser.cs" />
+    <Compile Include="Settings\SettingsPanel.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
