+++ InfiniteGoodsMod/Settings/SettingsPanel.cs
